+++ CHANGELOG.md
@@ -6,6 +6,28 @@ List all changes in various categories:
 * Bug fixes
 * Known issues
 
+## 2.0.0 (2018-10-09) ##
+
+### Release notes ###
+
+- OAuth2 token is now provided by default in Authorization header.
+- Auths does not take a list anymore but a var args instead.
+- MicrosoftOAuth2 renamed into AzureActiveDirectory.
+- Nonce is not a mandatory parameter anymore for MicrosoftOAuth2.
+
+### Enhancements ###
+
+- Authentication classes can now be imported from requests_auth.
+- JSONTokenFileCache can now be imported from requests_auth.
+- Okta authentication is now available (thanks to Sebastien De Menten).
+
+### Bug fixes ###
+
+- Oauth2 authentication was not working with Python 2.7
+- Update requests to 2.19.1 (latest version)
+- OAuth2 authentication now supports GET on token reception (thanks to Sebastien De Menten).
+- Extra parameters were not handled when using MicrosoftOAuth2 (now AzureActiveDirectory)
+
 ## 1.0.2 (2018-01-19) ##
 
 ### Release notes ###